dmvm installed

+++ run.sh
@@ -1,11 +1,18 @@
 #!/usr/bin/env bash
 
-echo "Выполняется резервное копирование данных..."
-
-cp $EMAIL_CRAWLER_HOME/data/crawler.sqlite $EMAIL_CRAWLER_HOME/backup/last-crawler.sqlite && \
-cp $EMAIL_CRAWLER_HOME/data/emails.csv $EMAIL_CRAWLER_HOME/backup/last-emails.csv && \
-cp $EMAIL_CRAWLER_HOME/link_for_crawler.txt $EMAIL_CRAWLER_HOME/backup/last-link_for_crawler.txt && \
-cp $EMAIL_CRAWLER_HOME/data/domains.csv $EMAIL_CRAWLER_HOME/backup/last-domains.csv && \
+echo "ЗАПУСКАЕМ СКРИП ИЗ ПАПКИ С КРАВЛЕРОМ !!!"
+export EMAIL_CRAWLER_HOME=$PWD
+{
+  if [ -f $EMAIL_CRAWLER_HOME/data/crawler.sqlite ]; then
+    echo "Выполняется резервное копирование данных..."
+    cp $EMAIL_CRAWLER_HOME/data/crawler.sqlite $EMAIL_CRAWLER_HOME/backup/last-crawler.sqlite && \
+    cp $EMAIL_CRAWLER_HOME/data/emails.csv $EMAIL_CRAWLER_HOME/backup/last-emails.csv && \
+    cp $EMAIL_CRAWLER_HOME/link_for_crawler.txt $EMAIL_CRAWLER_HOME/backup/last-link_for_crawler.txt && \
+    cp $EMAIL_CRAWLER_HOME/data/domains.csv $EMAIL_CRAWLER_HOME/backup/last-domains.csv
+  else
+    echo "Копировать нехер"
+  fi
+}
 
 echo "Запущен генератор ссылок..." && \
